Refactor ignore code into its own file
